# src/test/java/com/fieldServiceV2/assignment/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Technician","description":"Technician API"}],"paths":{"/api/technicians/{id}":{"get":{"tags":["Technician"],"summary":"Get a technician by ID","operationId":"getTechnicianById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Technician"}}}}}},"put":{"tags":["Technician"],"operationId":"updateTechnician","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Technician"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Technician"}}}}}},"delete":{"tags":["Technician"],"operationId":"deleteTechnician","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/jobs/{id}":{"get":{"tags":["job-order-controller"],"operationId":"getJobOrderById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobOrder"}}}}}},"put":{"tags":["job-order-controller"],"operationId":"updateJobOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobOrder"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobOrder"}}}}}}},"/api/jobs/{id}/status":{"put":{"tags":["job-order-controller"],"operationId":"updateJobOrderStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"status","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobOrder"}}}}}}},"/api/technicians":{"get":{"tags":["Technician"],"operationId":"getAllTechnicians","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Technician"}}}}}}},"post":{"tags":["Technician"],"operationId":"addTechnician","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Technician"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Technician"}}}}}}},"/api/jobs":{"get":{"tags":["job-order-controller"],"operationId":"getAllJobOrders","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JobOrder"}}}}}}},"post":{"tags":["job-order-controller"],"operationId":"addJobOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobOrder"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobOrder"}}}}}}},"/api/jobs/location":{"post":{"tags":["location-controller"],"operationId":"addLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Location"}}}}}}},"/api/reports/completed-jobs":{"get":{"tags":["reports-controller"],"operationId":"getCompletedJobsWithinTimeframe","parameters":[{"name":"start","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"end","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JobOrder"}}}}}}}},"/api/reports/active-jobs":{"get":{"tags":["reports-controller"],"operationId":"getActiveJobsWithinTimeframe","parameters":[{"name":"start","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"end","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JobOrder"}}}}}}}},"/api/jobs/{jobOrderId}/location":{"get":{"tags":["location-controller"],"operationId":"getJobLocation","parameters":[{"name":"jobOrderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}},"components":{"schemas":{"Technician":{"required":["email","name","phoneNumber"],"type":"object","properties":{"technicianId":{"type":"string","format":"uuid"},"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"maxLength":15,"minLength":10,"type":"string"}}},"JobOrder":{"required":["title"],"type":"object","properties":{"jobOrderId":{"type":"string","format":"uuid"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string","enum":["PENDING","IN_PROGRESS","CANCELLED","COMPLETED"]},"scheduledDate":{"type":"string","format":"date-time"},"technicianId":{"type":"string","format":"uuid"}}},"Location":{"required":["latitude","longitude"],"type":"object","properties":{"locationId":{"type":"string","format":"uuid"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"jobOrder":{"$ref":"#/components/schemas/JobOrder"}}}}}}